Address Toby's review on RandomStrategy NChooseK refactor
- Tighten NChooseKConstraint validator: feature must have bounds[0]==0
  or allow_zero=True; add invalid Domain spec.
- Move sample_valid_nchoosek_features from Domain to RandomStrategy as
  @staticmethod; remove legacy get_nchoosek_combinations and its tests.
- Add separate nchoosek_max_iters to RandomStrategy and pass it through.
- Drop redundant list() around con.features.
