+++ bofire/data_models/domain/domain.py
@@ -1,7 +1,4 @@
 import collections.abc
-import itertools
-import math
-import random
 import warnings
 from collections.abc import Sequence
 from typing import Any, Dict, Literal, Optional, Tuple, Union
@@ -16,7 +13,6 @@ from bofire.data_models.constraints.api import (
     Constraint,
     ConstraintNotFulfilledError,
     InterpointConstraint,
-    NChooseKConstraint,
 )
 from bofire.data_models.domain.constraints import Constraints
 from bofire.data_models.domain.features import Inputs, Outputs
@@ -134,181 +130,6 @@ class Domain(BaseModel):
         for c in self.constraints.get():
             c.validate_inputs(self.inputs)
         return self
-
-    # TODO: tidy this up
-    def get_nchoosek_combinations(self, exhaustive: bool = False):
-        """Get all possible NChooseK combinations
-
-        Args:
-            exhaustive (bool, optional): if True all combinations are returned. Defaults to False.
-
-        Returns:
-            Tuple(used_features_list, unused_features_list): used_features_list is a list of lists containing features used in each NChooseK combination.
-                unused_features_list is a list of lists containing features unused in each NChooseK combination.
-
-        """
-        if len(self.constraints.get(NChooseKConstraint)) == 0:
-            used_continuous_features = self.inputs.get_keys(ContinuousInput)
-            return used_continuous_features, []
-
-        used_features_list_all = []
-
-        # loops through each NChooseK constraint
-        for con in self.constraints.get(NChooseKConstraint):
-            assert isinstance(con, NChooseKConstraint)
-            used_features_list = []
-
-            if exhaustive:
-                for n in range(con.min_count, con.max_count + 1):
-                    used_features_list.extend(itertools.combinations(con.features, n))
-
-                if con.none_also_valid:
-                    used_features_list.append(())
-            else:
-                used_features_list.extend(
-                    itertools.combinations(con.features, con.max_count),
-                )
-
-            used_features_list_all.append(used_features_list)
-
-        used_features_list_all = list(
-            itertools.product(*used_features_list_all),
-        )  # product between NChooseK constraints
-
-        # format into a list of used features
-        used_features_list_formatted = []
-        for used_features_list in used_features_list_all:
-            used_features_list_flattened = [
-                item for sublist in used_features_list for item in sublist
-            ]
-            used_features_list_formatted.append(list(set(used_features_list_flattened)))
-
-        # sort lists
-        used_features_list_sorted = []
-        for used_features in used_features_list_formatted:
-            used_features_list_sorted.append(sorted(used_features))
-
-        # drop duplicates
-        used_features_list_no_dup = []
-        for used_features in used_features_list_sorted:
-            if used_features not in used_features_list_no_dup:
-                used_features_list_no_dup.append(used_features)
-
-        # remove combinations not fulfilling constraints
-        used_features_list_final = []
-        for combo in used_features_list_no_dup:
-            fulfil_constraints = []  # list of bools tracking if constraints are fulfilled
-            for con in self.constraints.get(NChooseKConstraint):
-                assert isinstance(con, NChooseKConstraint)
-                count = 0  # count of features in combo that are in con.features
-                for f in combo:
-                    if f in con.features:
-                        count += 1
-                if (
-                    count >= con.min_count
-                    and count <= con.max_count
-                    or count == 0
-                    and con.none_also_valid
-                ):
-                    fulfil_constraints.append(True)
-                else:
-                    fulfil_constraints.append(False)
-            if np.all(fulfil_constraints):
-                used_features_list_final.append(combo)
-
-        # features unused
-        features_in_cc = []
-        for con in self.constraints.get(NChooseKConstraint):
-            assert isinstance(con, NChooseKConstraint)
-            features_in_cc.extend(con.features)
-        features_in_cc = list(set(features_in_cc))
-        features_in_cc.sort()
-        unused_features_list = []
-        for used_features in used_features_list_final:
-            unused_features_list.append(
-                [f_key for f_key in features_in_cc if f_key not in used_features],
-            )
-
-        # postprocess
-        # used_features_list_final2 = []
-        # unused_features_list2 = []
-        # for used, unused in zip(used_features_list_final,unused_features_list):
-        #     if len(used) == 3:
-        #         used_features_list_final2.append(used), unused_features_list2.append(unused)
-
-        return used_features_list_final, unused_features_list
-
-    def sample_valid_nchoosek_features(
-        self,
-        seed: Optional[int] = None,
-        n: int = 1,
-        max_iters: int = 1000,
-    ) -> list[Tuple[str, ...]]:
-        """Sample sets of active feature keys uniformly from all valid subsets.
-
-        Includes (a) one group per ``NChooseKConstraint`` (respecting
-        ``min_count``, ``max_count``, and ``none_also_valid``) and (b) one
-        singleton group per ``ContinuousInput`` with ``allow_zero=True`` that
-        is not already part of any ``NChooseKConstraint``.
-
-        Within each group the subset size ``k`` is drawn with probability
-        proportional to ``C(n, k)`` and ``k`` features are then chosen
-        uniformly, so the per-group distribution is uniform over all valid
-        subsets. When ``NChooseKConstraint``s share features, the per-group
-        union may violate one of the constraints; in that case rejection
-        sampling is used (up to ``max_iters`` attempts per drawn combination).
-
-        Args:
-            seed: Random seed used to initialise the internal sampler.
-                Defaults to ``None`` (non-deterministic).
-            n: Number of combinations to draw. Defaults to 1.
-            max_iters: Maximum number of rejection-sampling attempts per
-                drawn combination. Defaults to 1000.
-
-        Returns:
-            A list of ``n`` sorted tuples of active feature keys.
-
-        Raises:
-            ValueError: If a valid combination is not found within
-                ``max_iters`` attempts.
-        """
-        rng = random.Random(seed)
-        groups: list[Tuple[list[str], list[int], list[int]]] = []
-        nchoosek_keys: set[str] = set()
-        nchoosek_cons = list(self.constraints.get(NChooseKConstraint))
-        for con in nchoosek_cons:
-            assert isinstance(con, NChooseKConstraint)
-            ks = list(range(con.min_count, con.max_count + 1))
-            if con.none_also_valid and 0 not in ks:
-                ks.insert(0, 0)
-            weights = [math.comb(len(con.features), k) for k in ks]
-            groups.append((list(con.features), ks, weights))
-            nchoosek_keys.update(con.features)
-        for feat in self.inputs.get(ContinuousInput):
-            assert isinstance(feat, ContinuousInput)
-            if feat.allow_zero and feat.key not in nchoosek_keys:
-                groups.append(([feat.key], [0, 1], [1, 1]))
-
-        results: list[Tuple[str, ...]] = []
-        for _ in range(n):
-            for _ in range(max_iters):
-                active: set[str] = set()
-                for features, ks, weights in groups:
-                    k = rng.choices(ks, weights=weights, k=1)[0]
-                    active.update(rng.sample(features, k))
-                if all(
-                    (con.none_also_valid and len(active & set(con.features)) == 0)
-                    or con.min_count <= len(active & set(con.features)) <= con.max_count
-                    for con in nchoosek_cons
-                ):
-                    results.append(tuple(sorted(active)))
-                    break
-            else:
-                raise ValueError(
-                    f"Failed to sample a valid NChooseK combination after "
-                    f"{max_iters} attempts.",
-                )
-        return results
 
     def coerce_invalids(self, experiments: pd.DataFrame) -> pd.DataFrame:
         """Coerces all invalid output measurements to np.nan
