+++ tests/bofire/data_models/domain/test_domain.py
@@ -507,75 +507,3 @@ def test_is_fulfilled():
         domain.is_fulfilled(experiments),
         pd.Series([True, False, False], index=experiments.index),
     )
-
-
-def test_sample_valid_nchoosek_features_uniform_over_subsets():
-    """With one NChooseK on n=5 features and k in [1, 3], there are
-    C(5,1)+C(5,2)+C(5,3) = 25 valid subsets. With uniform sampling each
-    should appear with frequency ~1/25.
-    """
-    inputs = [ContinuousInput(key=f"x{i}", bounds=(0, 1)) for i in range(5)]
-    constraint = NChooseKConstraint(
-        features=[f"x{i}" for i in range(5)],
-        min_count=1,
-        max_count=3,
-        none_also_valid=False,
-    )
-    domain = Domain(
-        inputs=Inputs(features=inputs),
-        constraints=Constraints(constraints=[constraint]),
-    )
-    n_samples = 25_000
-    samples = domain.sample_valid_nchoosek_features(seed=0, n=n_samples)
-    counts: dict = {}
-    for s in samples:
-        counts[s] = counts.get(s, 0) + 1
-    assert len(counts) == 25, f"Expected 25 unique subsets, got {len(counts)}"
-    expected = n_samples / 25
-    for subset, count in counts.items():
-        rel = abs(count - expected) / expected
-        assert (
-            rel < 0.20
-        ), f"Subset {subset} count {count} too far from expected {expected:.0f}"
-
-
-def test_sample_valid_nchoosek_features_none_also_valid():
-    """When none_also_valid=True, the empty subset is in the support."""
-    inputs = [ContinuousInput(key=f"x{i}", bounds=(0, 1)) for i in range(3)]
-    constraint = NChooseKConstraint(
-        features=["x0", "x1", "x2"],
-        min_count=2,
-        max_count=3,
-        none_also_valid=True,
-    )
-    domain = Domain(
-        inputs=Inputs(features=inputs),
-        constraints=Constraints(constraints=[constraint]),
-    )
-    samples = domain.sample_valid_nchoosek_features(seed=1, n=2000)
-    unique = set(samples)
-    # Valid subsets: () + C(3,2) + C(3,3) = 1 + 3 + 1 = 5
-    assert len(unique) == 5
-    assert () in unique
-
-
-def test_sample_valid_nchoosek_features_allow_zero_singletons():
-    """Without any NChooseK, allow_zero=True features form singleton groups."""
-    inputs = [
-        ContinuousInput(key="a", bounds=(0.1, 1), allow_zero=True),
-        ContinuousInput(key="b", bounds=(0.1, 1), allow_zero=True),
-        ContinuousInput(key="c", bounds=(0.1, 1)),
-    ]
-    domain = Domain(inputs=Inputs(features=inputs))
-    samples = domain.sample_valid_nchoosek_features(seed=2, n=2000)
-    unique = set(samples)
-    # Each of {a, b} can be on or off independently -> 4 subsets
-    assert unique == {(), ("a",), ("b",), ("a", "b")}
-
-
-def test_sample_valid_nchoosek_features_empty_returns_empty_tuple():
-    """Domain without NChooseK and without allow_zero features yields ()."""
-    inputs = [ContinuousInput(key="x", bounds=(0, 1))]
-    domain = Domain(inputs=Inputs(features=inputs))
-    samples = domain.sample_valid_nchoosek_features(seed=3, n=4)
-    assert samples == [(), (), (), ()]
